feat: :tada: configuracion de card html
feat: :construction: Done, just need to customize and maybe add transactions.

style: :lipstick: Hyperpersonalized

feat: :lipstick:  The radii of the boxes were adjusted to look very similar to those of the figma.

feat: :lipstick: Finish, was customized and made as similar as possible to that of the figma.

feat: :construction: finish