feat: :tada: configuracion del css de balance
feat: :construction: Moving forward, I'm almost done with the 3rd page and would proceed to customize

style: :lipstick: Hyperpersonalized

feat: :construction: The 3rd part is completed and is being customized.

feat: :lipstick: Finish, was customized and made as similar as possible to that of the figma.

feat: :construction: Some necessary adjustments were made, the trainer helped me to better adjust the page and improve it visually.