feat: :tada: Inicio de la configuracion del estilos index
feat: :construction: Avanzando ya casi por mitad

feat: :construction: I had to make adjustments to the HTML structure and modify the measurements in CSS

feat: :construction: Moving forward, I'm almost done with the 3rd page and would proceed to customize

feat: :construction: a border needs to be fixed and customized

style: :lipstick: personalize

style: :lipstick: Hyperpersonalized

feat: :construction: The 3rd part is completed and is being customized.

feat: :lipstick: Finish, was customized and made as similar as possible to that of the figma.

feat: :sparkles: I changed the settings a little bit and I can see a space in my cell phone, but in the simulator for the type of cell phone indicated by the teacher it looks very good.

feat: :construction: Some necessary adjustments were made, the trainer helped me to better adjust the page and improve it visually.

style: :lipstick: change icons